Removing @import "eyeglass/assets"
This was causing the Fatal error: Cannot read property 'eyeglass' of undefined

The reason was, with this line in place, that error was thrown in eyeglass_modules.js
in the canAccessModule function where it defines the canAccess variable.
That has an if statement that checks if branch.dependencies[name] is truthy.
Unfortunately with that "eyeglass/assets" import, "dependencies" is undefined
because the "branches" variable is equal to:
[ { name: 'spriting',
    version: '0.9.2',
    path: '/projects/js/eyeglass-spriting-grunt/node_modules/eyeglass-spriting' } ]

The 'name' variable in this case is 'eyeglass', hence the error message

Removing that @import statement allowed my sass to compile again